olajiAI/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "school-ai-gateway"
version = "0.1.0"
description = "AI-only FastAPI gateway that orchestrates intents and calls the SchoolOps Core API"
authors = [{name="SchoolOps", email="ops@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "redis>=5.0.4",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "tenacity>=8.2.3",
    "structlog>=24.1.0"
]
